moved generic macros of parsec into easy_parsec/
(and moved back parsec/base/ to parsec/)
