+++ include/cparsec3/parsec/ParsecRunner.h
@@ -1,5 +1,5 @@
 /* -*- coding: utf-8-unix -*- */
-#include "../../base/base_generics.h"
+#include "../base/base_generics.h"
 
 #include "parsec.h"
