+++ include/cparsec3/parsec/parser/ParsecPrim.h
@@ -3,7 +3,7 @@
 
 #include "../../base/base_generics.h"
 
-#include "../base/parsec.h"
+#include "../parsec.h"
 
 // -----------------------------------------------------------------------
 #define ParsecPrim(...) TYPE_NAME(ParsecPrim, __VA_ARGS__)
