+++ include/cparsec3/parsec/state.h
@@ -1,7 +1,7 @@
 /* -*- coding: utf-8-unix -*- */
 #pragma once
 
-#include "../../base/base.h"
+#include "../base/base.h"
 
 #include "sourcepos.h"
